#49 Reorganize subpackage structure
