+++ necst/core/auth/authorizer.py
@@ -7,7 +7,7 @@ from rclpy.callback_groups import MutuallyExclusiveCallbackGroup
 from rclpy.node import Node
 from std_srvs.srv import Empty
 
-from .. import config, namespace, utils
+from ... import config, namespace, utils
 from necst_msgs.srv import AuthoritySrv
